refactor: move fuelSign into utils
