+++ src/hooks/useSubmit/utils.ts
@@ -10,7 +10,7 @@ import {
 import { wagmiConfig } from "wagmiConfig"
 import { Chain, Chains, supportedChains } from "wagmiConfig/chains"
 import { DEFAULT_MESSAGE } from "./constants"
-import { MessageParams, SignProps, Validation } from "./types"
+import { FuelSignProps, MessageParams, SignProps, Validation } from "./types"
 
 export const signWithKeyPair = (keyPair: CryptoKeyPair, params: MessageParams) =>
   window.crypto.subtle
@@ -150,4 +150,27 @@ export const chainsOfAddressWithDeployedContract = async (
   localStorage.setItem(LOCALSTORAGE_KEY, JSON.stringify(res))
 
   return res
+}
+
+export const fuelSign = async ({
+  wallet,
+  address,
+  payload,
+  keyPair,
+  forcePrompt,
+  msg = DEFAULT_MESSAGE,
+  ts,
+}: FuelSignProps): Promise<[string, Validation]> => {
+  const params = createMessageParams(address, ts, msg, payload)
+  let sig = null
+
+  if (!!keyPair && !forcePrompt) {
+    params.method = ValidationMethod.KEYPAIR
+    sig = await signWithKeyPair(keyPair, params)
+  } else {
+    params.method = ValidationMethod.FUEL
+    sig = await wallet.signMessage(getMessage(params))
+  }
+
+  return [payload, { params, sig }]
 }
